requirements.txt
streamlit
segno
pillow
